+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/JackBurr/OldFiles/LeftAutoV5.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.JackBurr.Autonomous;
+package org.firstinspires.ftc.teamcode.JackBurr.OldFiles;
 
 //import com.acmerobotics.roadrunner.geometry.Pose2d;
 //import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -10,7 +10,6 @@ import com.acmerobotics.roadrunner.TrajectoryActionBuilder;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -24,10 +23,9 @@ import org.firstinspires.ftc.teamcode.JackBurr.Servos.DifferentialV2;
 import org.firstinspires.ftc.teamcode.JackBurr.Servos.GrippersV1;
 import org.firstinspires.ftc.teamcode.JackBurr.Servos.WristAxonV1;
 
-@Disabled
 @Config
 @Autonomous
-public class LeftAutoV4 extends LinearOpMode {
+public class LeftAutoV5 extends LinearOpMode {
     public DeliverySlidesV1 slides = new DeliverySlidesV1();
     public DeliveryAxonV1 deliveryAxonV1 = new DeliveryAxonV1();
     public RobotConstantsV1 robotConstantsV1 = new RobotConstantsV1();
@@ -49,19 +47,18 @@ public class LeftAutoV4 extends LinearOpMode {
     //----------------------------------------------------------------------------------------
     public static Pose2d startPose = new Pose2d(36, 62, Math.toRadians(-90)); // (60,0), 180
 
-    public static Vector2d position1 = new Vector2d(21, 42); // (40,-15) (x-20, y-15)
-    public static double position1HeadingDegrees = -120; // 150 (heading -30)
+    public static Vector2d position1 = new Vector2d(51, 47);
+    public static double position1HeadingDegrees = -135;
 
-    public static Vector2d position2 = new Vector2d(22, 57); // (46, -5) (x+6, y+10)
-    public static double position2HeadingDegrees = -120; //150 (heading +0)
+    public static Vector2d position2 = new Vector2d(47, 43);
+    public static double position2HeadingDegrees = -85;
 
-    public static double position3Degrees = -90; // 180 (heading +30)
+    public static Vector2d position4 = new Vector2d(61, 47);
 
-    public static double position4Degrees = -90; // 180 (heading +0)
+    public static Vector2d position5 = new Vector2d(47, 47);
+    public static Vector2d position6 = new Vector2d(47, 0);
+    public static Vector2d position7 = new Vector2d(10, 10);
 
-    public static Vector2d position5 = new Vector2d(17.75, 54.35); // (41.75, -7.65) (x - 4.25, y - 2.65)
-    public static Vector2d position6 = new Vector2d(30, 52); // (54, -10) (x + 12.25, y + 2.35)
-    public static double position6HeadingDegrees = -140; // 130 (heading - 50)
 
     //+X is left, +Y is backwards
 
@@ -100,24 +97,39 @@ public class LeftAutoV4 extends LinearOpMode {
         telemetry.addData(">", "Press Play to start op mode");
         telemetry.update();
 
-        Pose2d startPose = new Pose2d(60, 0, Math.toRadians(180));
 
         traj1Builder = drive.actionBuilder(startPose)
-                .splineTo(position1, Math.toRadians(position1HeadingDegrees));
+                .strafeTo(position1)
+                .turnTo(Math.toRadians(position1HeadingDegrees));
         traj2Builder = traj1Builder.fresh()
-                .lineToX(position2.x);
+                .strafeTo(position2)
+                .turnTo(Math.toRadians(position2HeadingDegrees))
+                .turnTo(Math.toRadians(position2HeadingDegrees))
+                .strafeTo(position2);
         traj3Builder = traj2Builder.fresh()
-                .turnTo(Math.toRadians(position3Degrees));
+                .strafeTo(position1)
+                .turnTo(Math.toRadians(position1HeadingDegrees));
         traj4Builder = traj3Builder.fresh()
-                .strafeTo(position5);
+                .turnTo(Math.toRadians(position2HeadingDegrees))
+                .strafeTo(position4);
+        traj5Builder = traj4Builder.fresh()
+                .strafeTo(position5)
+                .strafeTo(position6)
+                .turn(Math.toRadians(90))
+                .strafeTo(position7);
+
+        //traj3Builder = traj2Builder.fresh()
+                //.turnTo(Math.toRadians(position3Degrees));
+        //traj4Builder = traj3Builder.fresh()
+                //.strafeTo(position5);
         //traj5Builder = traj4Builder.fresh()
                 //.turnTo(Math.toRadians(position4Degrees));
-        traj6Builder = traj4Builder.fresh()
-                .strafeTo(position1)
-                .turnTo(Math.toRadians(position6HeadingDegrees));
-        traj7Builder = traj6Builder.fresh()
-                .strafeTo(position6)
-                .turnTo(Math.toRadians(position6HeadingDegrees));
+        //traj6Builder = traj4Builder.fresh()
+                //.strafeTo(position1)
+                //.turnTo(Math.toRadians(position6HeadingDegrees));
+        //traj7Builder = traj6Builder.fresh()
+                //.strafeTo(position6)
+                //.turnTo(Math.toRadians(position6HeadingDegrees));
 
 
 
@@ -127,9 +139,9 @@ public class LeftAutoV4 extends LinearOpMode {
         traj2 = traj2Builder.build();
         traj3 = traj3Builder.build();
         traj4 = traj4Builder.build();
-        //traj5 = traj5Builder.build();
-        traj6 = traj6Builder.build();
-        traj7 = traj7Builder.build();
+        traj5 = traj5Builder.build();
+        //traj6 = traj6Builder.build();
+        //traj7 = traj7Builder.build();
 
 
         // Trajectory traj2 = drive.trajectoryBuilder(traj1.end())
@@ -165,8 +177,8 @@ public class LeftAutoV4 extends LinearOpMode {
             if (step == 1){
                 while (deliveryTimer.seconds() < 2){
                     deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_GRAB);
-                    slides.runLeftSlideToPosition(robotConstantsV1.LEFT_SLIDE_HIGH_BASKET, 0.9);
-                    slides.runRightSlideToPosition(robotConstantsV1.RIGHT_SLIDE_HIGH_BASKET, 0.9);
+                    slides.runLeftSlideToPositionPID(robotConstantsV1.LEFT_SLIDE_HIGH_BASKET);
+                    slides.runRightSlideToPositionPID(robotConstantsV1.RIGHT_SLIDE_HIGH_BASKET);
                     if(!servoSet) {
                         deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_UP);
                         servoSet = true;
@@ -180,7 +192,6 @@ public class LeftAutoV4 extends LinearOpMode {
                         servoSet2 = true;
                     }
                     if(!traj4Followed) {
-                        Actions.runBlocking(traj2);
                         traj4Followed = true;
                     }
                 }
@@ -188,21 +199,24 @@ public class LeftAutoV4 extends LinearOpMode {
                     deliveryGrippers.setPosition(robotConstantsV1.DELIVERY_GRIPPERS_OPEN);
                 }
                 while (deliveryTimer.seconds() < 4.5) {
+                    intakeSlides.runToPosition(intakeTarget01, 1);
+                    diffV2.setTopRightServoPosition(robotConstantsV1.FRONT_RIGHT_HOVER);
+                    diffV2.setTopLeftServoPosition(robotConstantsV1.FRONT_LEFT_HOVER);
                     if (!servoSet) {
                         deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_GRAB);
                         grippers.setPosition(robotConstantsV1.GRIPPERS_OPEN);
                         servoSet = true;
                     }
                     if (slides.getLeftSlidePosition() != 0 || slides.getRightSlidePosition() != 0) {
-                        slides.runLeftSlideToPosition(0, 0.9);
-                        slides.runRightSlideToPosition(0, 0.9);
+                        slides.runLeftSlideToPositionPID(0);
+                        slides.runRightSlideToPositionPID(0);
                     } else if (slides.getRightSlidePosition() != 0 && slides.getLeftSlidePosition() != 0) {
-                        slides.runLeftSlideToPosition(0, 0.9);
-                        slides.runRightSlideToPosition(0, 0.9);
+                        slides.runLeftSlideToPositionPID(0);
+                        slides.runRightSlideToPositionPID(0);
                     }
                     if(!traj5Followed) {
-                        Actions.runBlocking(traj3);
-                        Actions.runBlocking(traj4);
+                        Actions.runBlocking(traj2);
+                        //Actions.runBlocking(traj4);
                         traj5Followed = true;
                     }
                 }
@@ -214,18 +228,15 @@ public class LeftAutoV4 extends LinearOpMode {
                 }
             }
             if(step == 2) {
-                while (intakeTimer.seconds() < 2){
-                    intakeSlides.runToPosition(intakeTarget01, 1);
+                while (intakeTimer.seconds() < 0.8){
+                    diffV2.setTopRightServoPosition(robotConstantsV1.FRONT_RIGHT_PICKUP);
+                    diffV2.setTopLeftServoPosition(robotConstantsV1.FRONT_LEFT_PICKUP);
                     if(!wristSet){
                         wrist.setPosition(robotConstantsV1.WRIST_CENTER);
                         wristSet = true;
                     }
                 }
-                while(intakeTimer.seconds() < 3){
-                    diffV2.setTopRightServoPosition(robotConstantsV1.FRONT_RIGHT_PICKUP);
-                    diffV2.setTopLeftServoPosition(robotConstantsV1.FRONT_LEFT_PICKUP);
-                }
-                while (intakeTimer.seconds() < 5.5){
+                while (intakeTimer.seconds() < 1.3){
                     grippers.setPosition(robotConstantsV1.GRIPPERS_GRAB);
                 }
                 step = 3;
@@ -244,17 +255,17 @@ public class LeftAutoV4 extends LinearOpMode {
                     diffV2.setTopLeftServoPosition(robotConstantsV1.FRONT_LEFT_TRANSFER);
                     diffV2.setTopRightServoPosition(robotConstantsV1.FRONT_RIGHT_TRANSFER);
                 }
-                while (intakeTimer.seconds() < 3.5){
+                while (intakeTimer.seconds() < 3){
                     intakeSlides.intakeAllTheWayIn();
                 }
-                while (intakeTimer.seconds() < 4.5){
+                while (intakeTimer.seconds() < 3.7){
                     deliveryGrippers.setPosition(robotConstantsV1.DELIVERY_GRIPPERS_CLOSE);
                 }
-                while (intakeTimer.seconds() < 5.2){
+                while (intakeTimer.seconds() < 4.5){
                     grippers.setPosition(robotConstantsV1.GRIPPERS_OPEN);
                 }
-                Actions.runBlocking(traj6);
-                Actions.runBlocking(traj7);
+                Actions.runBlocking(traj3);
+                //Actions.runBlocking(traj7);
                 deliveryTimer.reset();
                 servoSet = false;
                 servoSet2 = false;
@@ -262,44 +273,43 @@ public class LeftAutoV4 extends LinearOpMode {
                 step = 5;
             }
             if(step == 5){
-                while (deliveryTimer.seconds() < 2.5){
-                    deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_GRAB);
-                    slides.runLeftSlideToPosition(robotConstantsV1.LEFT_SLIDE_HIGH_BASKET, 0.9);
-                    slides.runRightSlideToPosition(robotConstantsV1.RIGHT_SLIDE_HIGH_BASKET, 0.9);
-                    if(!servoSet) {
+                while (deliveryTimer.seconds() < 2){
+                    slides.runLeftSlideToPositionPID(robotConstantsV1.LEFT_SLIDE_HIGH_BASKET);
+                    slides.runRightSlideToPositionPID(robotConstantsV1.RIGHT_SLIDE_HIGH_BASKET);
+                    if(deliveryTimer.seconds() > 1.5) {
                         deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_UP);
                         servoSet = true;
                     }
                     deliveryGrippers.setPosition(robotConstantsV1.DELIVERY_GRIPPERS_CLOSE);
                 }
-                while (deliveryTimer.seconds() < 4) {
-                    servoSet = false;
-                    if (!servoSet2) {
-                        deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_UP);
-                        servoSet2 = true;
-                    }
-                }
                 deliveryTimer.reset();
                 step = 6;
             }
             if(step == 6){
-                while (deliveryTimer.seconds() < 2) {
+                while (deliveryTimer.seconds() < 1) {
                     deliveryGrippers.setPosition(robotConstantsV1.DELIVERY_GRIPPERS_OPEN);
                 }
                 servoSet2 = false;
                 step = 7;
             }
             if(step == 7){
-                if(!servoSet2) {
-                    deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_GRAB);
+                while (deliveryTimer.seconds() < 3.5) {
+                    if (!servoSet2) {
+                        deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_GRAB);
+                    }
+                    if (slides.getLeftSlidePosition() != 0 || slides.getRightSlidePosition() != 0) {
+                        slides.runLeftSlideToPositionPID(0);
+                        slides.runRightSlideToPositionPID(0);
+                    } else if (slides.getRightSlidePosition() != 0 && slides.getLeftSlidePosition() != 0) {
+                        slides.runLeftSlideToPositionPID(0);
+                        slides.runRightSlideToPositionPID(0);
+                    }
                 }
-                if (slides.getLeftSlidePosition() != 0 || slides.getRightSlidePosition() != 0) {
-                    slides.runLeftSlideToPosition(0, 0.9);
-                    slides.runRightSlideToPosition(0, 0.9);
-                } else if (slides.getRightSlidePosition() != 0 && slides.getLeftSlidePosition() != 0) {
-                    slides.runLeftSlideToPosition(0, 0.9);
-                    slides.runRightSlideToPosition(0, 0.9);
-                }
+                step = 8;
+            }
+            if(step == 8){
+                Actions.runBlocking(traj5);
+                deliveryAxonV1.setPosition(robotConstantsV1.DELIVERY_LEVEL_ONE_ASCENT);
             }
         }
     }
