+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/JackBurr/OldFiles/OdoWheelsTest.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode.JackBurr.Autonomous;
+package org.firstinspires.ftc.teamcode.JackBurr.OldFiles;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
 
 public class OdoWheelsTest extends OpMode {
     @Override
