+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/JackBurr/OldFiles/RightAutoV6.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.JackBurr.Autonomous;
+package org.firstinspires.ftc.teamcode.JackBurr.OldFiles;
 
 import com.acmerobotics.roadrunner.Action;
 import com.acmerobotics.roadrunner.Pose2d;
